refactor(rimap-server): extract daemon run_with_shutdown for SCM reuse
